feat: I have added this to the system prompt If you believe after generating this file that another file not included must be edited you may make requests to change the plan by requesting to add files to the generation queue. I need you to make the changes necessary to give the AI the ability to fill out a new field and then add to the generation queue. Allow the AI to revisit files it has already visited
The task requires the AI to have the ability to dynamically add files to the generation queue and revisit previously processed files. This can be achieved by adding a new field to the `GeneratedCode` model, which the AI can populate. The `_execute_generation_loop` will then read this field and update its internal processing queue accordingly. The `html_utils.py` file needs to be updated to display a message to the user when the generation queue is dynamically updated, which is already handled by the `plan_updated` status in the `handleStatusUpdate` function.